refactor: Extract VertexAndName from OpenStreetMapModule
