+++ src/main/java/org/opentripplanner/graph_builder/module/osm/OpenStreetMapModule.java
@@ -12,7 +12,6 @@ import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.OptionalInt;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -1454,35 +1453,4 @@ public class OpenStreetMapModule implements GraphBuilderModule {
         }
     }
 
-    static class VertexAndName {
-
-        private final I18NString name;
-        private final OsmVertex vertex;
-
-        VertexAndName(I18NString name, OsmVertex vertex) {
-            this.name = name;
-            this.vertex = vertex;
-        }
-
-        public I18NString getName() {
-            return this.name;
-        }
-
-        public OsmVertex getVertex() {
-            return this.vertex;
-        }
-
-        @Override
-        public boolean equals(Object o) {
-            if (this == o) {return true;}
-            if (o == null || getClass() != o.getClass()) {return false;}
-            final VertexAndName that = (VertexAndName) o;
-            return Objects.equals(name, that.name) && vertex.equals(that.vertex);
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hash(name, vertex);
-        }
-    }
 }
